requirements.txt
Flask==3.0.0
psycopg2==2.9.9
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0
Flask-Cors==4.0.0
spotipy==2.23.0 